#81 the first commit for throttling config tab [vip:platform/pallas#860]
